Add setting for initializing in a disabled state (#12)
Fixes #10.

I was doing this in a local fork anyway to test an app, so I figured I
might as well open a PR.

This does what is described in
https://github.com/mbrea-c/bevy_firework/issues/10#issuecomment-2023331426.

Side note: It's not totally clear to me why `ParticleSpawnerData` is not
just included in `ParticleSpawnerBundle`. I experimented with removing
it here:
https://github.com/rparrett/bevy_firework/commit/d33aa0a5c0e7de70f6917bba224db88e53f78fb7
and at a glance, things don't seem broken.